修复一个溢出bug——当i>Integer.MAX_VALUE时时间会倒流 本次的问题行： CustomAnalyslsActivity.java:286 AnalysisFragment.java:298 影响内容：两个图表页柱状图横轴的时间显示

+++ app/src/main/java/com/bqmz001/moneynotes/fragment/AnalysisFragment.java
@@ -295,7 +295,7 @@ public class AnalysisFragment extends ViewPagerFragment {
                         int days = (int) ((stopTime + 1 - startTime) / 86400000);
                         for (int i = 0; i < days; i++) {
                             DailyNoteFakeCount dailyNoteFakeCount = new DailyNoteFakeCount();
-                            dailyNoteFakeCount.setTime(DateTimeUtil.timestampToDate(startTime + (i * 86400000)).substring(5, 10));
+                            dailyNoteFakeCount.setTime(DateTimeUtil.timestampToDate((startTime + ((long)i * 86400000))).substring(5, 10));
                             dailyNoteFakeCount.setStartTime(startTime + (i * 86400000));
                             dailyNoteFakeCount.setStopTime(startTime + ((i + 1) * 86400000) - 1);
                             dnfcs.add(dailyNoteFakeCount);
